// src/app/interfaces/movie.ts
export interface Movie {
    id?: number;
    budget?: string;
    originalLang?: string;
    originalTitle?: string;
    overview?: string;
    popularity?: string;
    releaseDate?: string;
    revenue?: string;
    runtime?: string;
    status?: string;
    tagline?: string;
    title?: string;
    voteAverage?: string;
    voteCount?: number;
    createdAt?: string;
    updatedAt?: string;
    deletedAt?: string;
}